Create Metro.h
Header file for the Metro class. This is the class where everything comes together.  This file will take two input files, one with a list of passengers and their info and another loading the lines of the DC Metro System, including the stops in each line. 

This class will hold all the metro system's lines in a single vector. It also can calculate earnings based on passengers and the lines (including the number of stops) they traveled through.

This function will also house the application's UI, executing various actions across all four classes.