Fniished Metro Class (MainMenu and CalcEarnings)
Defined the last 2 remaining functions for the Metro class, MainMenu and
CalcEarnings. This also marks the completion of the base version of this project

MainMenu: Displays avaialable executable actions and take user input to carry them out

CalcEarnings: Calulates the earnings of each line that day, taking into account
the number of passengers on each line, and the individual stops travelled by each
individual passenger. Also counts the total number of stops travelled that day.